@@ -15,69 +19,62 @@
     "@types/mocha": "^8.2.3",
     "argparse": "^1.0.10",
     "axios": "^0.21.1",
-    "chai": "^4.3.4",
+    "chai": "^4.3.10",
     "chai-as-promised": "^7.1.1",
     "chalk": "^4.1.0",
     "collections": "^5.1.12",
     "commander": "^8.3.0",
-    "ethereum-waffle": "^3.0.0",
+    "eslint": "^8.51.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "ethereum-waffle": "^4.0.10",
     "ethereumjs-abi": "^0.6.8",
     "ethers": "^5.7.0",
     "ethjs": "^0.4.0",
     "fs": "^0.0.1-security",
     "handlebars": "^4.7.6",
-    "hardhat": "=2.12.4",
+    "hardhat": "^2.18.3",
     "hardhat-contract-sizer": "^2.0.2",
     "hardhat-gas-reporter": "^1.0.9",
     "hardhat-typechain": "^0.3.3",
     "jsonwebtoken": "^8.5.1",
-    "merkletreejs": "^0.2.32",
+    "markdownlint-cli": "^0.33.0",
+    "merkletreejs": "^0.3.11",
     "mocha": "^9.0.2",
     "path": "^0.12.7",
-    "prettier": "^2.8.8",
-    "prettier-plugin-solidity": "^1.1.3",
     "querystring": "^0.2.0",
     "solc": "0.8.17",
     "solhint": "^3.6.2",
-    "solhint-plugin-prettier": "^0.0.5",
-    "solidity-coverage": "^0.8.2",
+    "solidity-coverage": "^0.8.5",
     "ts-generator": "^0.1.1",
     "ts-node": "^10.1.0",
     "typechain": "^4.0.0",
-    "typescript": "^4.3.5"
+    "typescript": "^4.6.4",
+    "zksync-ethers": "https://github.com/kelemeno/zksync-ethers#ethers-v5-feat/bridgehub"
   },
   "scripts": {
     "build": "hardhat compile",
     "clean": "hardhat clean",
-    "test": "CONTRACT_TESTS=1 yarn hardhat test test/unit_tests/*.spec.ts --network hardhat",
+    "clean:foundry": "forge clean",
+    "test": "hardhat test test/unit_tests/*.spec.ts --network hardhat",
     "test:foundry": "hardhat solpp && forge test",
-    "test:fork": "CONTRACT_TESTS=1 TEST_CONTRACTS_FORK=1 yarn run hardhat test test/unit_tests/*.fork.ts --network hardhat",
+    "test:fork": "TEST_CONTRACTS_FORK=1 yarn run hardhat test test/unit_tests/*.fork.ts --network hardhat",
     "coverage:foundry": "hardhat solpp && forge coverage",
-    "lint": "yarn lint:sol && yarn prettier:check",
-    "lint:sol-contracts": "solhint --max-warnings 40 contracts/**/*.sol",
-    "lint:sol-tests": "solhint --max-warnings 0 test/**/*.sol",
-    "lint:sol": "yarn lint:sol-contracts && yarn lint:sol-tests",
-    "prettier:check-contracts": "prettier --check contracts/**/*.sol",
-    "prettier:check-tests": "prettier --check test/**/*.sol",
-    "prettier:check": "yarn prettier:check-contracts && yarn prettier:check-tests",
-    "prettier:write-contracts": "prettier --write contracts/**/*.sol",
-    "prettier:write-tests": "prettier --write test/**/*.sol",
-    "prettier:write": "yarn prettier:write-contracts && yarn prettier:write-tests",
     "deploy-no-build": "ts-node scripts/deploy.ts",
+    "register-hyperchain": "ts-node scripts/register-hyperchain.ts",
     "deploy-weth-bridges": "ts-node scripts/deploy-weth-bridges.ts",
-    "initialize-weth-bridges": "ts-node scripts/initialize-weth-bridges.ts",
     "initialize-l2-weth-token": "ts-node scripts/initialize-l2-weth-token.ts",
-    "allow-list-manager": "ts-node scripts/allow-list-manager.ts",
     "deploy-erc20": "ts-node scripts/deploy-erc20.ts",
     "token-info": "ts-node scripts/token-info.ts",
     "deploy-testkit": "ts-node scripts/deploy-testkit.ts",
     "verify": "hardhat run --network env scripts/verify.ts",
-    "deploy-testnet-erc20": "ts-node scripts/deploy-testnet-token.ts",
     "read-variable": "ts-node scripts/read-variable.ts",
     "initialize-bridges": "ts-node scripts/initialize-bridges.ts",
-    "initialize-allow-list": "ts-node scripts/initialize-l1-allow-list.ts",
     "initialize-validator": "ts-node scripts/initialize-validator.ts",
     "initialize-governance": "ts-node scripts/initialize-governance.ts",
+    "migrate-governance": "ts-node scripts/migrate-governance.ts",
+    "display-governance": "ts-node scripts/display-governance.ts",
     "upgrade-1": "ts-node scripts/upgrades/upgrade-1.ts",
     "upgrade-2": "ts-node scripts/upgrades/upgrade-2.ts",
     "upgrade-3": "ts-node scripts/upgrades/upgrade-3.ts",
