@@ -1,5 +1,5 @@
 {
-  "name": "l2-zksync-contracts",
+  "name": "l2-contracts",
   "version": "0.1.0",
   "license": "MIT",
   "devDependencies": {
@@ -12,25 +12,23 @@
     "@nomiclabs/hardhat-ethers": "^2.0.0",
     "@nomiclabs/hardhat-etherscan": "^3.1.7",
     "@nomiclabs/hardhat-solpp": "^2.0.0",
-    "@openzeppelin/contracts": "4.9.2",
-    "@openzeppelin/contracts-upgradeable": "4.9.2",
+    "@openzeppelin/contracts": "4.9.5",
+    "@openzeppelin/contracts-upgradeable": "4.9.5",
     "@typechain/ethers-v5": "^2.0.0",
     "@types/chai": "^4.2.21",
     "@types/chai-as-promised": "^7.1.4",
     "@types/mocha": "^8.2.3",
-    "chai": "^4.3.4",
+    "chai": "^4.3.10",
     "chai-as-promised": "^7.1.1",
     "chalk": "^4.1.0",
     "commander": "^6.0.0",
     "ethers": "^5.7.0",
-    "hardhat": "=2.12.4",
+    "hardhat": "^2.18.3",
     "hardhat-typechain": "^0.3.3",
     "mocha": "^9.0.2",
-    "prettier": "^1.18.2",
-    "prettier-plugin-solidity": "=1.0.0-dev.22",
     "ts-node": "^10.1.0",
     "typechain": "^4.0.0",
-    "typescript": "^4.3.5",
+    "typescript": "^5.2.2",
     "zksync-web3": "^0.15.4"
   },
   "scripts": {
@@ -38,11 +36,16 @@
     "clean": "hardhat clean",
     "test": "hardhat test",
     "verify": "hardhat run src/verify.ts",
-    "deploy-testnet-paymaster": "ts-node src/deployTestnetPaymaster.ts",
-    "deploy-force-deploy-upgrader": "ts-node src/deployForceDeployUpgrader.ts",
+    "deploy-testnet-paymaster": "ts-node src/deploy-testnet-paymaster.ts",
+    "deploy-testnet-paymaster-through-l1": "ts-node src/deploy-testnet-paymaster-through-l1.ts",
+    "deploy-force-deploy-upgrader": "ts-node src/deploy-force-deploy-upgrader.ts",
+    "deploy-force-deploy-upgrader-through-l1": "ts-node src/deploy-force-deploy-upgrader-through-l1.ts",
+    "deploy-shared-bridge-on-l2": "ts-node scripts/deploy-shared-bridge-on-l2.ts",
+    "deploy-shared-bridge-on-l2-through-l1": "ts-node src/deploy-shared-bridge-on-l2-through-l1.ts",
     "publish-bridge-preimages": "ts-node src/publish-bridge-preimages.ts",
-    "deploy-l2-weth": "ts-node src/deployL2Weth.ts",
-    "upgrade-l2-erc20-contract": "ts-node src/upgradeL2BridgeImpl.ts"
+    "deploy-l2-weth": "ts-node src/deploy-l2-weth.ts",
+    "upgrade-bridge-contracts": "ts-node src/upgrade-bridge-impl.ts",
+    "update-l2-erc20-metadata": "ts-node src/update-l2-erc20-metadata.ts"
   },
   "dependencies": {
     "dotenv": "^16.0.3"
