@@ -1,58 +1,60 @@
-import { Command } from 'commander';
-import { Wallet } from 'ethers';
-import { Deployer } from '../src.ts/deploy';
-import { formatUnits, parseUnits } from 'ethers/lib/utils';
-import { web3Provider } from './utils';
-
-import * as fs from 'fs';
-import * as path from 'path';
+// hardhat import should be the first import in the file
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import * as hardhat from "hardhat";
+import { Command } from "commander";
+import { Wallet } from "ethers";
+import { Deployer } from "../src.ts/deploy";
+import { formatUnits, parseUnits } from "ethers/lib/utils";
+import { GAS_MULTIPLIER, web3Provider } from "./utils";
+import { ethTestConfig } from "../src.ts/utils";
 
 const provider = web3Provider();
-const testConfigPath = path.join(process.env.ZKSYNC_HOME as string, `etc/test_config/constant`);
-const ethTestConfig = JSON.parse(fs.readFileSync(`${testConfigPath}/eth.json`, { encoding: 'utf-8' }));
 
 async function main() {
-    const program = new Command();
+  const program = new Command();
 
-    program
-        .option('--private-key <private-key>')
-        .option('--gas-price <gas-price>')
-        .option('--nonce <nonce>')
-        .action(async (cmd) => {
-            const deployWallet = cmd.privateKey
-                ? new Wallet(cmd.privateKey, provider)
-                : Wallet.fromMnemonic(
-                      process.env.MNEMONIC ? process.env.MNEMONIC : ethTestConfig.mnemonic,
-                      "m/44'/60'/0'/0/1"
-                  ).connect(provider);
-            console.log(`Using deployer wallet: ${deployWallet.address}`);
+  program
+    .option("--private-key <private-key>")
+    .option("--chain-id <chain-id>")
+    .option("--gas-price <gas-price>")
+    .option("--nonce <nonce>")
+    .action(async (cmd) => {
+      const deployWallet = cmd.privateKey
+        ? new Wallet(cmd.privateKey, provider)
+        : Wallet.fromMnemonic(
+            process.env.MNEMONIC ? process.env.MNEMONIC : ethTestConfig.mnemonic,
+            "m/44'/60'/0'/0/1"
+          ).connect(provider);
+      console.log(`Using deployer wallet: ${deployWallet.address}`);
 
-            const gasPrice = cmd.gasPrice ? parseUnits(cmd.gasPrice, 'gwei') : await provider.getGasPrice();
-            console.log(`Using gas price: ${formatUnits(gasPrice, 'gwei')} gwei`);
+      const gasPrice = cmd.gasPrice
+        ? parseUnits(cmd.gasPrice, "gwei")
+        : (await provider.getGasPrice()).mul(GAS_MULTIPLIER);
+      console.log(`Using gas price: ${formatUnits(gasPrice, "gwei")} gwei`);
 
-            const nonce = cmd.nonce ? parseInt(cmd.nonce) : await deployWallet.getTransactionCount();
-            console.log(`Using nonce: ${nonce}`);
+      const nonce = cmd.nonce ? parseInt(cmd.nonce) : await deployWallet.getTransactionCount();
+      console.log(`Using nonce: ${nonce}`);
 
-            const deployer = new Deployer({
-                deployWallet,
-                verbose: true
-            });
+      const deployer = new Deployer({
+        deployWallet,
+        verbose: true,
+      });
 
-            const zkSync = deployer.zkSyncContract(deployWallet);
-            const validatorTimelock = deployer.validatorTimelock(deployWallet);
-            const tx = await zkSync.setValidator(validatorTimelock.address, true);
-            console.log(`Transaction sent with hash ${tx.hash} and nonce ${tx.nonce}`);
-            const receipt = await tx.wait();
+      const zkSync = deployer.stateTransitionContract(deployWallet);
+      const validatorTimelock = deployer.validatorTimelock(deployWallet);
+      const tx = await zkSync.setValidator(validatorTimelock.address, true);
+      console.log(`Transaction sent with hash ${tx.hash} and nonce ${tx.nonce}`);
+      const receipt = await tx.wait();
 
-            console.log(`Validator is set, gasUsed: ${receipt.gasUsed.toString()}`);
-        });
+      console.log(`Validator is set, gasUsed: ${receipt.gasUsed.toString()}`);
+    });
 
-    await program.parseAsync(process.argv);
+  await program.parseAsync(process.argv);
 }
 
 main()
-    .then(() => process.exit(0))
-    .catch((err) => {
-        console.error('Error:', err);
-        process.exit(1);
-    });
+  .then(() => process.exit(0))
+  .catch((err) => {
+    console.error("Error:", err);
+    process.exit(1);
+  });
