@@ -1,241 +1,164 @@
-import { expect } from 'chai';
-import { ethers } from 'ethers';
-import * as hardhat from 'hardhat';
-import { REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT } from 'zksync-web3/build/src/utils';
-import { IZkSync, IZkSyncFactory } from 'zksync-web3/build/typechain';
-import { Action, diamondCut, facetCut } from '../../src.ts/diamondCut';
-import {
-    AllowList,
-    AllowListFactory,
-    DiamondInitFactory,
-    GettersFacetFactory,
-    MailboxFacetFactory,
-    TestnetERC20Token,
-    TestnetERC20TokenFactory
-} from '../../typechain';
-import { IL1Bridge } from '../../typechain/IL1Bridge';
-import { IL1BridgeFactory } from '../../typechain/IL1BridgeFactory';
-import { AccessMode, getCallRevertReason } from './utils';
+import { expect } from "chai";
+import { ethers, Wallet } from "ethers";
+import * as hardhat from "hardhat";
+import { Interface } from "ethers/lib/utils";
 
-describe(`L1ERC20Bridge tests`, function () {
-    let owner: ethers.Signer;
-    let randomSigner: ethers.Signer;
-    let allowList: AllowList;
-    let l1ERC20Bridge: IL1Bridge;
-    let erc20TestToken: TestnetERC20Token;
-    let testnetERC20TokenContract: ethers.Contract;
-    let l1Erc20BridgeContract: ethers.Contract;
-    let zksyncContract: IZkSync;
+import type { Bridgehub, L1SharedBridge } from "../../typechain";
+import { L1SharedBridgeFactory, BridgehubFactory, TestnetERC20TokenFactory } from "../../typechain";
+import type { IL1ERC20Bridge } from "../../typechain/IL1ERC20Bridge";
+import { IL1ERC20BridgeFactory } from "../../typechain/IL1ERC20BridgeFactory";
 
-    before(async () => {
-        [owner, randomSigner] = await hardhat.ethers.getSigners();
+import { ADDRESS_ONE, ethTestConfig } from "../../src.ts/utils";
+import { getTokens } from "../../src.ts/deploy-token";
+import type { Deployer } from "../../src.ts/deploy";
+import { initialTestnetDeploymentProcess } from "../../src.ts/deploy-test-process";
 
-        const gettersFactory = await hardhat.ethers.getContractFactory(`GettersFacet`);
-        const gettersContract = await gettersFactory.deploy();
-        const gettersFacet = GettersFacetFactory.connect(gettersContract.address, gettersContract.signer);
+import { depositERC20, getCallRevertReason } from "./utils";
 
-        const mailboxFactory = await hardhat.ethers.getContractFactory('MailboxFacet');
-        const mailboxContract = await mailboxFactory.deploy();
-        const mailboxFacet = MailboxFacetFactory.connect(mailboxContract.address, mailboxContract.signer);
+describe("L1ERC20Bridge tests", function () {
+  let owner: ethers.Signer;
+  let randomSigner: ethers.Signer;
+  let deployWallet: Wallet;
+  let deployer: Deployer;
+  let l1ERC20BridgeAddress: string;
+  let l1ERC20Bridge: IL1ERC20Bridge;
+  let sharedBridgeProxy: L1SharedBridge;
+  let erc20TestToken: ethers.Contract;
+  let bridgehub: Bridgehub;
+  let chainId = "9"; // Hardhat config ERA_CHAIN_ID
+  const functionSignature = "0x11a2ccc1";
 
-        const allowListFactory = await hardhat.ethers.getContractFactory('AllowList');
-        const allowListContract = await allowListFactory.deploy(await allowListFactory.signer.getAddress());
-        allowList = AllowListFactory.connect(allowListContract.address, allowListContract.signer);
+  before(async () => {
+    [owner, randomSigner] = await hardhat.ethers.getSigners();
 
-        const diamondInitFactory = await hardhat.ethers.getContractFactory('DiamondInit');
-        const diamondInitContract = await diamondInitFactory.deploy();
-        const diamondInit = DiamondInitFactory.connect(diamondInitContract.address, diamondInitContract.signer);
+    const gasPrice = await owner.provider.getGasPrice();
 
-        const dummyHash = new Uint8Array(32);
-        dummyHash.set([1, 0, 0, 1]);
-        const dummyAddress = ethers.utils.hexlify(ethers.utils.randomBytes(20));
-        const diamondInitData = diamondInit.interface.encodeFunctionData('initialize', [
-            {
-                verifier: dummyAddress,
-                governor: await owner.getAddress(),
-                admin: await owner.getAddress(),
-                genesisBatchHash: ethers.constants.HashZero,
-                genesisIndexRepeatedStorageChanges: 0,
-                genesisBatchCommitment: ethers.constants.HashZero,
-                allowList: allowList.address,
-                verifierParams: {
-                    recursionCircuitsSetVksHash: ethers.constants.HashZero,
-                    recursionLeafLevelVkHash: ethers.constants.HashZero,
-                    recursionNodeLevelVkHash: ethers.constants.HashZero
-                },
-                zkPorterIsAvailable: false,
-                l2BootloaderBytecodeHash: dummyHash,
-                l2DefaultAccountBytecodeHash: dummyHash,
-                priorityTxMaxGasLimit: 10000000,
-            }
-        ]);
+    deployWallet = Wallet.fromMnemonic(ethTestConfig.test_mnemonic3, "m/44'/60'/0'/0/1").connect(owner.provider);
+    const ownerAddress = await deployWallet.getAddress();
+    process.env.ETH_CLIENT_CHAIN_ID = (await deployWallet.getChainId()).toString();
 
-        const facetCuts = [
-            facetCut(gettersFacet.address, gettersFacet.interface, Action.Add, false),
-            facetCut(mailboxFacet.address, mailboxFacet.interface, Action.Add, true)
-        ];
+    const tx = {
+      from: owner.getAddress(),
+      to: deployWallet.address,
+      value: ethers.utils.parseEther("1000"),
+      nonce: owner.getTransactionCount(),
+      gasLimit: 100000,
+      gasPrice: gasPrice,
+    };
 
-        const diamondCutData = diamondCut(facetCuts, diamondInit.address, diamondInitData);
+    await owner.sendTransaction(tx);
 
-        const diamondProxyFactory = await hardhat.ethers.getContractFactory('DiamondProxy');
-        const chainId = hardhat.network.config.chainId;
-        const diamondProxyContract = await diamondProxyFactory.deploy(chainId, diamondCutData);
+    process.env.CHAIN_ETH_ZKSYNC_NETWORK_ID = "9"; // the legacy functions work only for ERA, which has a specific chainId
+    deployer = await initialTestnetDeploymentProcess(deployWallet, ownerAddress, gasPrice, []);
+    chainId = deployer.chainId.toString();
 
-        const l1Erc20BridgeFactory = await hardhat.ethers.getContractFactory('L1ERC20Bridge');
-        l1Erc20BridgeContract = await l1Erc20BridgeFactory.deploy(
-            diamondProxyContract.address,
-            allowListContract.address
-        );
-        l1ERC20Bridge = IL1BridgeFactory.connect(l1Erc20BridgeContract.address, l1Erc20BridgeContract.signer);
+    bridgehub = BridgehubFactory.connect(deployer.addresses.Bridgehub.BridgehubProxy, deployWallet);
 
-        const testnetERC20TokenFactory = await hardhat.ethers.getContractFactory('TestnetERC20Token');
-        testnetERC20TokenContract = await testnetERC20TokenFactory.deploy('TestToken', 'TT', 18);
-        erc20TestToken = TestnetERC20TokenFactory.connect(
-            testnetERC20TokenContract.address,
-            testnetERC20TokenContract.signer
-        );
+    l1ERC20BridgeAddress = deployer.addresses.Bridges.ERC20BridgeProxy;
 
-        await erc20TestToken.mint(await randomSigner.getAddress(), ethers.utils.parseUnits('10000', 18));
-        await erc20TestToken
-            .connect(randomSigner)
-            .approve(l1Erc20BridgeContract.address, ethers.utils.parseUnits('10000', 18));
+    l1ERC20Bridge = IL1ERC20BridgeFactory.connect(l1ERC20BridgeAddress, deployWallet);
+    sharedBridgeProxy = L1SharedBridgeFactory.connect(deployer.addresses.Bridges.SharedBridgeProxy, deployWallet);
 
-        await (await allowList.setAccessMode(diamondProxyContract.address, AccessMode.Public)).wait();
+    const tokens = getTokens();
+    const tokenAddress = tokens.find((token: { symbol: string }) => token.symbol == "DAI")!.address;
+    erc20TestToken = TestnetERC20TokenFactory.connect(tokenAddress, owner);
 
-        // Exposing the methods of IZkSync to the diamond proxy
-        zksyncContract = IZkSyncFactory.connect(diamondProxyContract.address, diamondProxyContract.provider);
-    });
+    await erc20TestToken.mint(await randomSigner.getAddress(), ethers.utils.parseUnits("10000", 18));
+    await erc20TestToken.connect(randomSigner).approve(l1ERC20BridgeAddress, ethers.utils.parseUnits("10000", 18));
+  });
 
-    it(`Should not allow an un-whitelisted address to deposit`, async () => {
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge
-                .connect(randomSigner)
-                .deposit(
-                    await randomSigner.getAddress(),
-                    testnetERC20TokenContract.address,
-                    0,
-                    0,
-                    0,
-                    ethers.constants.AddressZero
-                )
-        );
-        expect(revertReason).equal(`nr`);
+  it("Check should initialize through governance", async () => {
+    const l1SharedBridgeInterface = new Interface(hardhat.artifacts.readArtifactSync("L1SharedBridge").abi);
+    const upgradeCall = l1SharedBridgeInterface.encodeFunctionData("initializeChainGovernance(uint256,address)", [
+      chainId,
+      ADDRESS_ONE,
+    ]);
 
-        await (await allowList.setAccessMode(l1Erc20BridgeContract.address, AccessMode.Public)).wait();
-    });
+    const txHash = await deployer.executeUpgrade(sharedBridgeProxy.address, 0, upgradeCall);
 
-    it(`Should not allow depositing zero amount`, async () => {
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge
-                .connect(randomSigner)
-                .deposit(
-                    await randomSigner.getAddress(),
-                    testnetERC20TokenContract.address,
-                    0,
-                    0,
-                    0,
-                    ethers.constants.AddressZero
-                )
-        );
-        expect(revertReason).equal(`2T`);
-    });
+    expect(txHash).not.equal(ethers.constants.HashZero);
+  });
 
-    it(`Should deposit successfully`, async () => {
-        const depositorAddress = await randomSigner.getAddress();
-        await depositERC20(
-            l1ERC20Bridge.connect(randomSigner),
-            zksyncContract,
-            depositorAddress,
-            testnetERC20TokenContract.address,
-            ethers.utils.parseUnits('800', 18),
-            10000000
-        );
-    });
-
-    it(`Should revert on finalizing a withdrawal with wrong message length`, async () => {
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(0, 0, 0, '0x', [])
-        );
-        expect(revertReason).equal(`kk`);
-    });
-
-    it(`Should revert on finalizing a withdrawal with wrong function signature`, async () => {
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(0, 0, 0, ethers.utils.randomBytes(76), [])
-        );
-        expect(revertReason).equal(`nt`);
-    });
-
-    it(`Should revert on finalizing a withdrawal with wrong batch number`, async () => {
-        const functionSignature = `0x11a2ccc1`;
-        const l1Receiver = await randomSigner.getAddress();
-        const l2ToL1message = ethers.utils.hexConcat([
-            functionSignature,
-            l1Receiver,
-            testnetERC20TokenContract.address,
-            ethers.constants.HashZero
-        ]);
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(10, 0, 0, l2ToL1message, [])
-        );
-        expect(revertReason).equal(`xx`);
-    });
-
-    it(`Should revert on finalizing a withdrawal with wrong length of proof`, async () => {
-        const functionSignature = `0x11a2ccc1`;
-        const l1Receiver = await randomSigner.getAddress();
-        const l2ToL1message = ethers.utils.hexConcat([
-            functionSignature,
-            l1Receiver,
-            testnetERC20TokenContract.address,
-            ethers.constants.HashZero
-        ]);
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(0, 0, 0, l2ToL1message, [])
-        );
-        expect(revertReason).equal(`xc`);
-    });
-
-    it(`Should revert on finalizing a withdrawal with wrong proof`, async () => {
-        const functionSignature = `0x11a2ccc1`;
-        const l1Receiver = await randomSigner.getAddress();
-        const l2ToL1message = ethers.utils.hexConcat([
-            functionSignature,
-            l1Receiver,
-            testnetERC20TokenContract.address,
-            ethers.constants.HashZero
-        ]);
-        const revertReason = await getCallRevertReason(
-            l1ERC20Bridge
-                .connect(randomSigner)
-                .finalizeWithdrawal(0, 0, 0, l2ToL1message, Array(9).fill(ethers.constants.HashZero))
-        );
-        expect(revertReason).equal(`nq`);
-    });
-});
-
-async function depositERC20(
-    bridge: IL1Bridge,
-    zksyncContract: IZkSync,
-    l2Receiver: string,
-    l1Token: string,
-    amount: ethers.BigNumber,
-    l2GasLimit: number,
-    l2RefundRecipient = ethers.constants.AddressZero
-) {
-    const gasPrice = await bridge.provider.getGasPrice();
-    const gasPerPubdata = REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT;
-    const neededValue = await zksyncContract.l2TransactionBaseCost(gasPrice, l2GasLimit, gasPerPubdata);
-
-    await bridge.deposit(
-        l2Receiver,
-        l1Token,
-        amount,
-        l2GasLimit,
-        REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT,
-        l2RefundRecipient,
-        {
-            value: neededValue
-        }
+  it("Should not allow depositing zero amount", async () => {
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge.connect(randomSigner)[
+        // solhint-disable-next-line no-unexpected-multiline
+        "deposit(address,address,uint256,uint256,uint256,address)"
+      ](await randomSigner.getAddress(), erc20TestToken.address, 0, 0, 0, ethers.constants.AddressZero)
     );
-}
+    expect(revertReason).equal("0T");
+  });
+
+  it("Should deposit successfully", async () => {
+    const depositorAddress = await randomSigner.getAddress();
+    await depositERC20(
+      l1ERC20Bridge.connect(randomSigner),
+      bridgehub,
+      chainId,
+      depositorAddress,
+      erc20TestToken.address,
+      ethers.utils.parseUnits("800", 18),
+      10000000
+    );
+  });
+
+  it("Should revert on finalizing a withdrawal with wrong message length", async () => {
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(0, 0, 0, "0x", [ethers.constants.HashZero])
+    );
+    expect(revertReason).equal("ShB wrong msg len");
+  });
+
+  it("Should revert on finalizing a withdrawal with wrong function signature", async () => {
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge
+        .connect(randomSigner)
+        .finalizeWithdrawal(0, 0, 0, ethers.utils.randomBytes(76), [ethers.constants.HashZero])
+    );
+    expect(revertReason).equal("ShB Incorrect message function selector");
+  });
+
+  it("Should revert on finalizing a withdrawal with wrong batch number", async () => {
+    const l1Receiver = await randomSigner.getAddress();
+    const l2ToL1message = ethers.utils.hexConcat([
+      functionSignature,
+      l1Receiver,
+      erc20TestToken.address,
+      ethers.constants.HashZero,
+    ]);
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(10, 0, 0, l2ToL1message, [])
+    );
+    expect(revertReason).equal("xx");
+  });
+
+  it("Should revert on finalizing a withdrawal with wrong length of proof", async () => {
+    const l1Receiver = await randomSigner.getAddress();
+    const l2ToL1message = ethers.utils.hexConcat([
+      functionSignature,
+      l1Receiver,
+      erc20TestToken.address,
+      ethers.constants.HashZero,
+    ]);
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge.connect(randomSigner).finalizeWithdrawal(0, 0, 0, l2ToL1message, [])
+    );
+    expect(revertReason).equal("xc");
+  });
+
+  it("Should revert on finalizing a withdrawal with wrong proof", async () => {
+    const l1Receiver = await randomSigner.getAddress();
+    const l2ToL1message = ethers.utils.hexConcat([
+      functionSignature,
+      l1Receiver,
+      erc20TestToken.address,
+      ethers.constants.HashZero,
+    ]);
+    const revertReason = await getCallRevertReason(
+      l1ERC20Bridge
+        .connect(randomSigner)
+        .finalizeWithdrawal(0, 0, 0, l2ToL1message, Array(9).fill(ethers.constants.HashZero))
+    );
+    expect(revertReason).equal("ShB withd w proof");
+  });
+});
