@@ -1,99 +1,48 @@
-import * as hre from 'hardhat';
-import * as fs from 'fs';
-import { exec as _exec, spawn as _spawn } from 'child_process';
+import type { CompilerPaths } from "./utils";
+import { spawn, compilerLocation, prepareCompilerPaths } from "./utils";
+import * as fs from "fs";
+import { Command } from "commander";
 
-import { getZksolcUrl, saltFromUrl } from '@matterlabs/hardhat-zksync-solc';
-import { getCompilersDir } from 'hardhat/internal/util/global-dir';
-import path from 'path';
-
-const COMPILER_VERSION = '1.3.14';
+const COMPILER_VERSION = "1.3.18";
 const IS_COMPILER_PRE_RELEASE = false;
 
-async function compilerLocation(): Promise<string> {
-    const compilersCache = await getCompilersDir();
-
-    let salt = '';
-
-    if (IS_COMPILER_PRE_RELEASE) {
-        const url = getZksolcUrl('https://github.com/matter-labs/zksolc-prerelease', hre.config.zksolc.version);
-        salt = saltFromUrl(url);
-    }
-
-    return path.join(compilersCache, 'zksolc', `zksolc-v${COMPILER_VERSION}${salt ? '-' : ''}${salt}`);
-}
-
-// executes a command in a new shell
-// but pipes data to parent's stdout/stderr
-export function spawn(command: string) {
-    command = command.replace(/\n/g, ' ');
-    const child = _spawn(command, { stdio: 'inherit', shell: true });
-    return new Promise((resolve, reject) => {
-        child.on('error', reject);
-        child.on('close', (code) => {
-            code == 0 ? resolve(code) : reject(`Child process exited with code ${code}`);
-        });
-    });
-}
-
-export async function compileYul(path: string, files: string[], outputDirName: string | null) {
-    if (!files.length) {
-        console.log(`No test files provided in folder ${path}.`);
-        return;
-    }
-    let paths = preparePaths(path, files, outputDirName);
-
-    const zksolcLocation = await compilerLocation();
-    await spawn(
-        `${zksolcLocation} ${paths.absolutePathSources}/${paths.outputDir} --optimization 3 --system-mode --yul --bin --overwrite -o ${paths.absolutePathArtifacts}/${paths.outputDir}`
-    );
+export async function compileYul(paths: CompilerPaths, file: string) {
+  const zksolcLocation = await compilerLocation(COMPILER_VERSION, IS_COMPILER_PRE_RELEASE);
+  await spawn(
+    `${zksolcLocation} ${paths.absolutePathSources}/${file} --optimization 3 --system-mode --yul --bin --overwrite -o ${paths.absolutePathArtifacts}`
+  );
 }
 
 export async function compileYulFolder(path: string) {
-    let files: string[] = (await fs.promises.readdir(path)).filter((fn) => fn.endsWith('.yul'));
-    for (const file of files) {
-        await compileYul(path, [file], `${file}`);
-    }
+  const paths = prepareCompilerPaths(path);
+  const files: string[] = (await fs.promises.readdir(path)).filter((fn) => fn.endsWith(".yul"));
+  for (const file of files) {
+    await compileYul(paths, `${file}`);
+  }
 }
-
-
-function preparePaths(path: string, files: string[], outputDirName: string | null): CompilerPaths {
-    const filePaths = files
-        .map((val, _) => {
-            return `sources/${val}`;
-        })
-        .join(' ');
-    const outputDir = outputDirName || files[0];
-    // This script is located in `system-contracts/scripts`, so we get one directory back.
-    const absolutePathSources = `${__dirname}/../${path}`; 
-    const absolutePathArtifacts = `${__dirname}/../${path}/artifacts`;
-
-    return new CompilerPaths(filePaths, outputDir, absolutePathSources, absolutePathArtifacts);
-}
-
-class CompilerPaths {
-    public filePath: string;
-    public outputDir: string;
-    public absolutePathSources: string;
-    public absolutePathArtifacts: string;
-    constructor(filePath: string, outputDir: string, absolutePathSources: string, absolutePathArtifacts: string) {
-        this.filePath = filePath;
-        this.outputDir = outputDir;
-        this.absolutePathSources = absolutePathSources;
-        this.absolutePathArtifacts = absolutePathArtifacts;
-    }
-}
-
 
 async function main() {
-    await compileYulFolder('contracts');
-    await compileYulFolder('contracts/precompiles');
-    await compileYulFolder('bootloader/build');
-    await compileYulFolder('bootloader/tests');
+  const program = new Command();
+
+  program.version("0.1.0").name("compile yul").description("publish preimages for the L2 contracts");
+
+  program.command("compile-bootloader").action(async () => {
+    await compileYulFolder("bootloader/build");
+    await compileYulFolder("bootloader/tests");
+  });
+
+  program.command("compile-precompiles").action(async () => {
+    await compileYulFolder("contracts-preprocessed");
+    await compileYulFolder("contracts-preprocessed/precompiles");
+    await compileYulFolder("contracts-preprocessed/precompiles/test-contracts");
+  });
+
+  await program.parseAsync(process.argv);
 }
 
 main()
-    .then(() => process.exit(0))
-    .catch((err) => {
-        console.error('Error:', err.message || err);
-        process.exit(1);
-    });
+  .then(() => process.exit(0))
+  .catch((err) => {
+    console.error("Error:", err.message || err);
+    process.exit(1);
+  });
